EOL-01 - refactor pakeges, add gradle file
