+++ build.gradle
@@ -1,0 +1,20 @@
+wrapper {
+    distributionType = Wrapper.DistributionType.ALL
+}
+
+/*test {
+    useJUnitPlatform()
+}*/
+
+/*ext {
+    branchDefault = "master"
+    lombokVersion="1.18.12"
+
+}*/
+
+dependencies {
+
+    implementation "org.apache.logging.log4j:log4j-to-slf4j"
+    implementation "org.apache.logging.log4j:log4j-slf4j-impl"
+
+}
