+++ src/main/java/edge/of/luck/fxml/ChooseEnemiesWindow.fxml
