--- src/main/java/entities/GameResult.java
@@ -1,12 +1,0 @@
-package main.java.entities;
-
-public enum GameResult {
-    WIN("Вы угадали"),
-    LOSE("Вы не угадали");
-
-    String result;
-
-    public String getResult() { return result; }
-
-    GameResult(String result) { this.result = result; }
-}
